Sorted out project
