+++ EasyInvestBank/Bank/admin.py
@@ -1,3 +1,5 @@
 from django.contrib import admin
-
+from .models import Transaction_List
 # Register your models here.
+
+admin.site.register(Transaction_List)
