+++ EasyInvestBank/Bank/forms.py
@@ -1,13 +1,33 @@
 from django import forms
-from .models import Transaction
+from .models import Transaction_List
 
 
-class MoneyTransferForm (forms.ModelForm):
+
+# class MoneyTransferForm (forms.ModelForm):
+#     class Meta:
+#         model = Transaction
+#         fields = [
+#             "senders_name",
+#             "recipients_name",
+#             "recipients_account_number", 
+#             "amount"
+#         ]
+        
+        
+        
+class MoneyTransferForm(forms.ModelForm):
+    amount = forms.DecimalField(max_digits=60, decimal_places=2, label='Amount')
+    receiver_IBAN = forms.CharField(max_length=100)
+    sender_IBAN = forms.CharField(max_length=100)
+    sender_FULLNAME = forms.CharField(max_length=100)
+    
     class Meta:
-        model = Transaction
+        model = Transaction_List
         fields = [
-            "senders_name",
-            "recipients_name",
-            "recipients_account_number", 
-            "amount"
+            'amount',
+            'receiver_IBAN',
+            'sender_IBAN',
+            'sender_FULLNAME'
         ]
+        
+
