--- EasyInvestBank/Bank/templates/money_transfer.html
@@ -1,29 +1,0 @@
-{% extends 'base.html' %}
-       
-{% block content %}
-    <section id="welcome_post" class="col-9">
-        <h2>Welcome, <span id="prof_name">{{ user }}.</span> </h2>
-        <h4 style="font-weight: 300;">It's good to have you here.</h4>   
-
-        <form action="m{ url 'money_transfer' }" method="POST">
-            {% csrf_token %}
-
-            <h3 style="font-weight: 300;">
-                {{ form }}
-            </h3>
-            
-            <small>By transferring money from your account, you agree to the <a href="#">Terms and Conditions</a> of amount transfer</small>. <br>
-            <button type="submit" class="btn btn-danger">Transfer amount</button>
-        </form>
-    </section>
-{% endblock %}
-
-{% block scriptcontent %}
-    <script>
-        $(".current").removeClass("current");
-        $("#transfer").addClass("current");
-
-        $("input").addClass("form-control");
-        $("label").prepend("<br>")
-    </script>
-{% endblock %}
