--- EasyInvestBank/User/json
@@ -1,17 +1,0 @@
-{
-'status': 'ok',
-'totalResults': 70,
-'articles':
-    [
-        {'source':
-            {'id': 'google-news', 'name': 'Google News'},
-        'author': None,
-        'title': "Man drove his family all night to make his brother's wedding after American Airlines booked them on a flight to a different country then canceled their replacement - Yahoo! Voices",
-        'description': None,
-        'url': 'https://news.google.com/__i/rss/rd/articles/CBMiSWh0dHBzOi8vd3d3LnlhaG9vLmNvbS9uZXdzL21hbi1kcm92ZS1mYW1pbHktbmlnaHQtYnJvdGhlcnMtMTEwNTEwNDkxLmh0bWzSAVJodHRwczovL25ld3MueWFob28uY29tL2FtcGh0bWwvbmV3cy9tYW4tZHJvdmUtZmFtaWx5LW5pZ2h0LWJyb3RoZXJzLTExMDUxMDQ5MS5odG1s?oc=5',
-        'urlToImage': None,
-        'publishedAt': '2022-08-21T11:05:10Z',
-        'content': None
-        }
-    ]
-}
