+++ EasyInvestBank/User/models.py
@@ -4,9 +4,7 @@ from django.db import models
 from django.contrib.auth.models import (
     BaseUserManager, AbstractBaseUser
 )
-import requests
 import random
-from django.utils.timezone import now
 from .constants import COUNTRY_CHOICES
 
 
@@ -102,11 +100,5 @@ class Account(AbstractBaseUser):
     
     
     
-class Transaction_List(models.Model):
-    
-    sender_IBAN = models.CharField(max_length=100, default=0)
-    sender_FULLNAME = models.CharField(max_length=100, default=0)
-    amount = models.DecimalField(max_digits=60, decimal_places=2)
-    receiver_IBAN = models.CharField(max_length=100)
-    sent_at = models.DateTimeField(auto_now_add=True)
+
