+++ EasyInvestBank/User/templates/User/base.html
@@ -45,9 +45,8 @@
         </div>
         </div>
     </nav>
-
-{% block content %}
-{% endblock %}
+    {% block content %}
+    {% endblock %}
 
   </body>
 </html>
