--- EasyInvestBank/news/templates/news.html
@@ -1,14 +1,0 @@
-
-{% extends "base.html" %}
-
-{% block content %}
-
-  <a href="{{ url }}">{{ title }}</a>
-</br>
-  <img src="{{ image }}" alt="alternatetext" width="300" height="300">
-<br />
-<br />
-  <a href="{{ url2 }}">{{ title2 }}</a>
-</br>
-  <img src="{{ image2 }}" alt="alternatetext" width="300" height="300">
-{% endblock %}
